Improved: naming for login utils and used common loader
