+++ .env.example
@@ -15,3 +15,4 @@ VUE_APP_WEBHOOK_ENUMS={"NEW_PRODUCTS":"products/create","DELETE_PRODUCTS":"produ
 VUE_APP_PERMISSION_ID=
 VUE_APP_ALIAS=
 VUE_APP_DEFAULT_LOG_LEVEL="error"
+VUE_APP_LOGIN_URL="http://launchpad.hotwax.io"
