+++ src/main.ts
@@ -5,8 +5,7 @@ import './registerServiceWorker'
 import { DateTime } from 'luxon';
 import logger from './logger';
 
-
-import { IonicVue, loadingController } from '@ionic/vue';
+import { IonicVue } from '@ionic/vue';
 
 /* Core CSS required for Ionic components to work properly */
 import '@ionic/vue/css/core.css';
@@ -28,34 +27,13 @@ import '@ionic/vue/css/display.css';
 import './theme/variables.css';
 import '@hotwax/apps-theme';
 
-import i18n, { translate } from './i18n'
+import i18n from './i18n'
 import store from './store'
 import permissionPlugin from '@/authorization';
 import permissionRules from '@/authorization/Rules';
 import permissionActions from '@/authorization/Actions';
 import { dxpComponents } from 'dxp-components'
-import { getAndSetUserDetails, getUserTokenAndOms, confirmSessionEnd, logout } from './user-utils';
-
-const loader = {
-  value: null as any,
-  present: async () => {
-    if (!loader.value) {
-      loader.value = await loadingController
-        .create({
-          message: translate("Logging in"),
-          translucent: false,
-          backdropDismiss: false
-        });
-    }
-    loader.value.present();
-  },
-  dismiss: () => {
-    if (loader.value) {
-      loader.value.dismiss();
-      loader.value = null as any;
-    }
-  }
-}
+import { login, getUserTokenAndOms, confirmSessionEnd, logout, loader } from './user-utils';
 
 const app = createApp(App)
   .use(IonicVue, {
@@ -72,7 +50,7 @@ const app = createApp(App)
     actions: permissionActions
   })
   .use(dxpComponents, {
-    getAndSetUserDetails,
+    login,
     getUserTokenAndOms,
     confirmSessionEnd,
     logout,
