+++ src/user-utils/index.ts
@@ -1,8 +1,9 @@
 import { translate } from '@/i18n'
 import store from '@/store'
 import { alertController } from '@ionic/core'
+import { loadingController } from '@ionic/vue'
 
-const getAndSetUserDetails = async (payload: any) => store.dispatch('user/getAndSetUserDetails', payload)
+const login = async (payload: any) => store.dispatch('user/login', payload)
 
 const getUserTokenAndOms = async () => {
   return {
@@ -32,8 +33,30 @@ const confirmSessionEnd = (appOms: string) => {
 
 const logout = async () => store.dispatch('user/logout')
 
+const loader = {
+  value: null as any,
+  present: async (message: string) => {
+    if (!loader.value) {
+      loader.value = await loadingController
+        .create({
+          message: translate(message),
+          translucent: false,
+          backdropDismiss: false
+        });
+    }
+    loader.value.present();
+  },
+  dismiss: () => {
+    if (loader.value) {
+      loader.value.dismiss();
+      loader.value = null as any;
+    }
+  }
+}
+
 export {
-  getAndSetUserDetails,
+  login,
+  loader,
   getUserTokenAndOms,
   confirmSessionEnd,
   logout
